Introducing cacheable events
Events can be cached through

    send_event(name, data, :cached => true)

They will be published like any other event, but will also be saved in the
Redis key-value store. Later, when a new dashing application starts running,
all the cached events will arrive right when /events is loaded.

This is specially useful for widgets which are not frequently updated.