Fix events connections never closing
The thing is, redis.psusbcribe was blocking the stream, and since we sometimes
have no recurring events, the connections were never getting closed. This
commit fixes that by spawning two threads, one to listen to and stream events,
and other to tick the *connection is still open check* each 5s.

Also, we release the database connection before executing the controller
action, so if there are too many live streams open, the pool of database
connections won't exhaust.